Feat(Rideshare - Backend) : Implement Rate entity relationship to user and driver class

+++ Rideshare.Application/Features/Rates/Handlers/DeleteRateCommandHandler.cs
@@ -29,20 +29,32 @@ namespace Rideshare.Application.Features.Rates.Handlers
 			
 			var response = new BaseResponse<Unit>();
 
-			var rate = await _unitOfWork.RateRepository.Get(request.RateId);
+			var rate = await _unitOfWork.RateRepository.Get(request.Id);
 
 			if (rate == null)
 				throw new NotFoundException("Resource Not Found");
+			
+			// Check if the UserId in the request matches the UserId of the rate in the database
+			if (rate.UserId != request.UserId)
+				throw new UnauthorizedAccessException("You are not authorized to delete this rate.");
+
 
 
 			if (await _unitOfWork.RateRepository.Delete(rate) == 0)
 				throw new InternalServerErrorException("Database Error: Unable To Save");
 			
-			// var driver = await _unitOfWork.DriverRepository.Get(rate.DriverId);
-			// driver.Rate[0] -= rate.Rate;
-			// driver.Rate[1] -= 1;
-			// if  (await _unitOfWork.DriverRepository.Update(driver) == 0)
-			//     throw new InternalServerErrorException("Database Error: Unable To Save");
+			var driver = await _unitOfWork.DriverRepository.Get(rate.DriverId);
+			driver.Rate[0] -= rate.Rate;
+			driver.Rate[1] -= 1;
+			
+			double total, count;
+			total = driver.Rate[0];
+			count = driver.Rate[1];
+			double average = (total / count);
+			driver.Rate[2] = average;
+			
+			if  (await _unitOfWork.DriverRepository.Update(driver) == 0)
+				throw new InternalServerErrorException("Database Error: Unable To Save");
 
 			response.Success = true;
 			response.Message = "Deletion Succeeded";
